refactor(webviews)!: rename src/webviews/api/ to src/webviews/webviewIntegration/
The folder is the local glue between this extension and
@microsoft/vscode-ext-react-webview; "api" was too generic and overloaded
with the extension's public API, REST/data API, and the package API.
webviewIntegration accurately describes the role and the contents
(extension-host controller, router, browser-side hook). All consumer imports
are updated. Skill docs (webview-trpc-messaging, react-webview-architecture)
are updated to match. Breaking change for in-tree consumers; npm package
surface is unchanged.
