+++ .github/skills/webview-trpc-messaging/SKILL.md
@@ -21,15 +21,15 @@ useTrpcClient() hook                      WebviewController
 
 **Key files** (read as needed for implementation details):
 
-| File                                                     | Purpose                                                                                    |
-| -------------------------------------------------------- | ------------------------------------------------------------------------------------------ |
-| `@microsoft/vscode-ext-react-webview/server` (trpc)            | tRPC init, `publicProcedure`, `createMiddleware`, `router`, `BaseRouterContext`            |
-| `src/webviews/api/trpc.ts`                              | DocumentDB-specific telemetry middleware + `publicProcedureWithTelemetry` + `WithTelemetry` |
-| `src/webviews/api/appRouter.ts`                         | Root router bundling all view routers + `commonRouter` + DocumentDB `BaseRouterContext`    |
-| `@microsoft/vscode-ext-react-webview/server` (WebviewController) | WebviewPanel lifecycle, tRPC message dispatcher (queries, mutations, subscriptions, abort) |
-| `src/webviews/api/WebviewController.ts`                 | DocumentDB-tuned wrapper that pre-fills router + bundle layout                              |
-| `src/webviews/api/useTrpcClient.ts`                     | React hook providing the tRPC client (pre-typed against `AppRouter`)                        |
-| `@microsoft/vscode-ext-react-webview` (vscodeLink)            | Custom tRPC link bridging `postMessage` transport                                          |
+| File                                                              | Purpose                                                                                                              |
+| ----------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------- |
+| `@microsoft/vscode-ext-react-webview/server` (trpc)               | tRPC init, `publicProcedure`, `createMiddleware`, `router`, `BaseRouterContext`                                      |
+| `src/webviews/webviewIntegration/appRouter.ts`                    | Root router + telemetry middleware + `publicProcedureWithTelemetry` + `WithTelemetry` + DocumentDB `BaseRouterContext` |
+| `src/webviews/webviewIntegration/configuration.ts`                | Consumer-owned knobs (telemetry namespace, bundle layout, dev-server host)                                           |
+| `@microsoft/vscode-ext-react-webview/server` (WebviewController)  | WebviewPanel lifecycle, tRPC message dispatcher (queries, mutations, subscriptions, abort)                           |
+| `src/webviews/webviewIntegration/WebviewControllerBase.ts`        | DocumentDB-tuned base class that pre-fills router + bundle layout                                                    |
+| `src/webviews/webviewIntegration/useTrpcClient.ts`                | React hook providing the tRPC client (pre-typed against `AppRouter`)                                                 |
+| `@microsoft/vscode-ext-react-webview` (vscodeLink)                | Custom tRPC link bridging `postMessage` transport                                                                    |
 
 ## Creating a New Router
 
@@ -41,7 +41,7 @@ Extend `BaseRouterContext` with view-specific fields:
 
 ```typescript
 // src/webviews/documentdb/myView/myViewRouter.ts
-import { type BaseRouterContext } from '../../api/appRouter';
+import { type BaseRouterContext } from '../../webviewIntegration/appRouter';
 
 export type RouterContext = BaseRouterContext & {
   clusterId: string;
@@ -55,7 +55,7 @@ export type RouterContext = BaseRouterContext & {
 
 ```typescript
 import { z } from 'zod';
-import { publicProcedureWithTelemetry, router, type WithTelemetry } from '../../api/trpc';
+import { publicProcedureWithTelemetry, router, type WithTelemetry } from '../../webviewIntegration/appRouter';
 
 export const myViewRouter = router({
   // Query with telemetry (preferred for operations that touch external services)
@@ -76,7 +76,7 @@ export const myViewRouter = router({
 ### 3. Register in appRouter
 
 ```typescript
-// src/webviews/api/appRouter.ts
+// src/webviews/webviewIntegration/appRouter.ts
 import { myViewRouter } from '../../documentdb/myView/myViewRouter';
 
 export const appRouter = router({
@@ -93,10 +93,10 @@ export const appRouter = router({
 
 ```typescript
 // src/webviews/documentdb/myView/myViewController.ts
-import { WebviewController } from '../../api/WebviewController';
+import { WebviewControllerBase } from '../../webviewIntegration/WebviewControllerBase';
 import { type RouterContext } from './myViewRouter';
 
-export class MyViewController extends WebviewController<MyViewConfig> {
+export class MyViewController extends WebviewControllerBase<MyViewConfig> {
   constructor(initialData: MyViewConfig) {
     super(ext.context, title, 'myViewName', initialData);
 
@@ -113,14 +113,14 @@ export class MyViewController extends WebviewController<MyViewConfig> {
 }
 ```
 
-> **Important:** The `webviewName` in the `WebviewController` constructor is the **registry key** (must match a key in `WebviewRegistry`, e.g. `collectionView`). The `webviewName` in the tRPC context is a **telemetry label** used in telemetry event names (e.g. `collectionView`). These may be the same string but serve different purposes — do not confuse them.
+> **Important:** The `webviewName` in the `WebviewControllerBase` constructor is the **registry key** (must match a key in `WebviewRegistry`, e.g. `collectionView`). The `webviewName` in the tRPC context is a **telemetry label** used in telemetry event names (e.g. `collectionView`). These may be the same string but serve different purposes — do not confuse them.
 
 ### 5. Register in WebviewRegistry
 
-Add your React component to the registry. The key must match the `webviewName` passed to `WebviewController`'s constructor. The `WebviewName` type (exported from the same file) ensures compile-time validation of webview names.
+Add your React component to the registry. The key must match the `webviewName` passed to `WebviewControllerBase`'s constructor. The `WebviewName` type (exported from the same file) ensures compile-time validation of webview names.
 
 ```typescript
-// src/webviews/api/WebviewRegistry.ts
+// src/webviews/webviewIntegration/WebviewRegistry.ts
 import { MyView } from '../../documentdb/myView/MyView';
 
 export const WebviewRegistry = {
@@ -240,7 +240,7 @@ sub.unsubscribe();
 ## Client-Side Hook Usage
 
 ```tsx
-import { useTrpcClient } from '../api/useTrpcClient';
+import { useTrpcClient } from '../webviewIntegration/useTrpcClient';
 import { useConfiguration } from '@microsoft/vscode-ext-react-webview';
 
 export const MyComponent = () => {
