+++ src/webviews/documentdb/collectionView/components/toolbar/ToolbarMainView.tsx
@@ -37,7 +37,7 @@ import {
 import { useConfiguration } from '@microsoft/vscode-ext-react-webview';
 import * as l10n from '@vscode/l10n';
 import { useContext, type JSX } from 'react';
-import { useTrpcClient } from '../../../../api/useTrpcClient';
+import { useTrpcClient } from '../../../../webviewIntegration/useTrpcClient';
 import { CollectionViewContext } from '../../collectionViewContext';
 import { type CollectionViewWebviewConfigurationType } from '../../collectionViewController';
 import { useHideScrollbarsDuringResize } from '../../hooks/useHideScrollbarsDuringResize';
