+++ src/webviews/documentdb/collectionView/components/toolbar/ToolbarViewNavigation.tsx
@@ -8,7 +8,7 @@ import { ArrowLeftFilled, ArrowPreviousFilled, ArrowRightFilled } from '@fluentu
 import * as l10n from '@vscode/l10n';
 import { useContext } from 'react';
 import { UsageImpact } from '../../../../../utils/surveyTypes';
-import { useTrpcClient } from '../../../../api/useTrpcClient';
+import { useTrpcClient } from '../../../../webviewIntegration/useTrpcClient';
 import { CollectionViewContext } from '../../collectionViewContext';
 import { ToolbarDividerTransparent } from './ToolbarDividerTransparent';
