+++ src/webviews/documentdb/documentView/documentsViewController.ts
@@ -6,7 +6,7 @@
 import * as vscode from 'vscode';
 import { API } from '../../../DocumentDBExperiences';
 import { ext } from '../../../extensionVariables';
-import { WebviewControllerBase } from '../../api/WebviewControllerBase';
+import { WebviewControllerBase } from '../../webviewIntegration/WebviewControllerBase';
 import { type RouterContext } from './documentsViewRouter';
 
 export type DocumentsViewWebviewConfigurationType = {
