+++ src/webviews/index.tsx
@@ -9,8 +9,8 @@ import { type l10nJsonFormat } from '@vscode/l10n';
 import type * as React from 'react';
 import { createRoot } from 'react-dom/client'; // eslint-disable-line import/no-internal-modules
 import { type WebviewApi } from 'vscode-webview';
-import { type WebviewName, WebviewRegistry } from './api/WebviewRegistry';
 import { DynamicThemeProvider } from './theme/DynamicThemeProvider';
+import { type WebviewName, WebviewRegistry } from './webviewIntegration/WebviewRegistry';
 
 export type ViewKey = WebviewName;
